+++ exe/arduino_ci.rb
@@ -9,7 +9,6 @@ VAR_CUSTOM_INIT_SCRIPT = "CUSTOM_INIT_SCRIPT".freeze
 VAR_USE_SUBDIR         = "USE_SUBDIR".freeze
 VAR_EXPECT_EXAMPLES    = "EXPECT_EXAMPLES".freeze
 VAR_EXPECT_UNITTESTS   = "EXPECT_UNITTESTS".freeze
-VAR_SKIP_LIBPROPS      = "SKIP_LIBRARY_PROPERTIES".freeze
 
 @failure_count = 0
 @passfail = proc { |result| result ? "✓" : "✗" }
@@ -22,7 +21,6 @@ class Parser
     output_options = {
       skip_unittests: false,
       skip_compilation: false,
-      skip_library_properties: false,
       ci_config: {
         "unittest" => unit_config
       },
@@ -36,10 +34,6 @@ class Parser
       end
 
       opts.on("--skip-examples-compilation", "Don't compile example sketches") do |p|
-        output_options[:skip_compilation] = p
-      end
-
-      opts.on("--skip-library-properties", "Don't validate library.properties entries") do |p|
         output_options[:skip_compilation] = p
       end
 
@@ -388,33 +382,6 @@ def choose_platform_set(config, reason, desired_platforms, library_properties)
   end
 end
 
-# tests of sane library.properties values
-def perform_property_tests(cpp_library)
-  phase("library.properties validation")
-  return inform("Skipping library.properties tests") { "as requested via command line" } if @cli_options[:skip_library_properties]
-  return inform("Skipping library.properties tests") { "as requested via environment" } unless ENV[VAR_SKIP_LIBPROPS].nil?
-  return inform("Skipping library.properties tests") { "file not found" } unless cpp_library.library_properties?
-
-  props = cpp_library.library_properties
-
-  props.depends&.each do |l|
-    assure("library.properties 'depends=' entry '#{l}' is available via the library manager") { @backend.library_available?(l) }
-  end
-
-  # the IDE would add these entries to a sketch (as "#include <...>" lines), they are nothing to do with the compioler
-  props.includes&.map(&:strip)&.map(&Pathname::method(:new))&.each do |f|
-    if (cpp_library.path + f).exist?
-      inform("library.properties 'includes=' entry found") { f }
-    elsif (cpp_library.path + "src" + f).exist?
-      inform("library.properties 'includes=' entry found") { Pathname.new("src") + f }
-    else
-      # this is if they want to "#include <math>" or something -- may or may not be valid!  so just warn.
-      warn("library.properties 'includes=' entry '#{f}' does not refer to a file in the library")
-    end
-  end
-
-end
-
 # Unit test procedure
 def perform_unit_tests(cpp_library, file_config)
   phase("Unit testing")
@@ -559,8 +526,6 @@ else
     false
   end
 end
-
-perform_property_tests(cpp_library)
 
 install_arduino_library_dependencies(
   cpp_library.arduino_library_dependencies,